Apply discount when the customer's total revenue is above EUR 1000
Set up a strategy to calculate discounts, so that it is easily
extendable with new types of discounts in the future.

Implemented 1 type of discount: high total revenue.

This is a discount where customer's that have already bought for
more than EUR 1000 in previous orders get a discount of 10% on the
whole order.